requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
playwright==1.54.0
trafilatura==1.12.2
beautifulsoup4==4.13.4
sentence-transformers==5.1.0
faiss-cpu==1.12.0
numpy==1.26.4
google-generativeai==0.8.5
httpx==0.28.1
